♻️ refactored code
