+++ server/githubsrm/apis/__init__.py
@@ -1,9 +1,7 @@
-from .aws import BotoService
 from .utils import check_token
 from .models import Entry
 from .checks_models import EntryCheck
-from .throttle import PostThrottle
 
-service = BotoService()
+
 open_entry_checks = EntryCheck()
 open_entry = Entry()
