+++ server/githubsrm/maintainer/models.py
@@ -6,9 +6,7 @@ import pymongo
 from django.conf import settings
 from administrator import jwt_keys
 
-from apis.aws import BotoService
-
-service = BotoService()
+from core import service
 
 
 class Entry:
