+++ server/tests/test_admin.py
@@ -265,6 +265,7 @@ class TestClient(unittest.TestCase):
                 "Content-type": "application/json", "X-RECAPTCHA-TOKEN": "TestToken",
                 "Authorization": f"Bearer {admin_jwt}"
             }, params={"role": "project"})
+        print(response.json())
         self.assertEqual(response.status_code, 200)
         self.clean()
 
@@ -311,6 +312,7 @@ class TestClient(unittest.TestCase):
                 "Content-type": "application/json", "X-RECAPTCHA-TOKEN": "TestToken",
                 "Authorization": f"Bearer {admin_jwt}"
             }, params={"role": "project"})
+        print(response.json())
         self.assertEqual(response.status_code, 200)
 
         response = self.client.post(
